backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
cryptography==44.0.2
defusedxml==0.7.1
Django==5.1.6
django-debug-toolbar==5.0.1
django-filter==25.1
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.0
drf-extra-fields==3.7.0
filetype==1.2.0
idna==3.10
oauthlib==3.2.2
pillow==11.1.0
psycopg2==2.9.7
pycparser==2.22
PyJWT==2.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
reportlab==4.3.1
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
